02-the-12-leverage-points/tac-2/app/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Natural Language SQL Interface Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.13",
    "uvicorn==0.34.3",
    "python-multipart==0.0.20",
    "openai==1.88.0",
    "anthropic==0.54.0",
    "pandas==2.3.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
]
